+++ superwall/src/main/java/com/superwall/sdk/store/AutomaticPurchaseController.kt
@@ -303,6 +303,7 @@ class AutomaticPurchaseController(
             } else {
                 SubscriptionStatus.Inactive
             }
+        Superwall.instance.setSubscriptionStatus(status)
 
         if (!Superwall.initialized) {
             Logger.debug(